+++ oauth-proxy/oauthHandlers/tokenHandlerStrategyClasses/saveDocumentStrategies/saveDocumentNoStrategy.js
@@ -1,8 +1,6 @@
 class SaveDocumentNoStrategy {
-  constructor() {}
   // eslint-disable-next-line no-unused-vars
   async saveDocumentToDynamo(document, tokens) {
-    return null;
   }
 }
