+++ oauth-proxy/oauthHandlers/tokenHandlerStrategyClasses/tokenHandlerClientBuilder.js
@@ -32,7 +32,7 @@ const {
 const {
   GetPatientInfoFromLaunchStrategy,
 } = require("./getPatientInfoStrategies/getPatientInfoFromLaunchStrategy");
-
+const { parseBasicAuth } = require("../../utils");
 const buildTokenHandlerClient = (
   redirect_uri,
   issuer,
@@ -58,7 +58,7 @@ const buildTokenHandlerClient = (
     validateToken
   );
   return new TokenHandlerClient(
-    strategies.getTokenResponseStrategy,
+    strategies.tokenHandlerStrategy,
     strategies.pullDocumentFromDynamoStrategy,
     strategies.saveDocumentToDynamoStrategy,
     strategies.getPatientInfoStrategy,
@@ -76,7 +76,6 @@ const getStrategies = (
   dynamoClient,
   config,
   req,
-  clientMetadata,
   validateToken
 ) => {
   let strategies;
@@ -85,7 +84,7 @@ const getStrategies = (
       tokenHandlerStrategy: new RefreshTokenStrategy(
         req,
         logger,
-        new issuer.Client(clientMetadata)
+        new issuer.Client(createClientMetadata(redirect_uri, req, config))
       ),
       pullDocumentFromDynamoStrategy: new PullDocumentByRefreshTokenStrategy(
         req,
@@ -111,7 +110,7 @@ const getStrategies = (
         req,
         logger,
         redirect_uri,
-        new issuer.Client(clientMetadata)
+        new issuer.Client(createClientMetadata(redirect_uri, req, config))
       ),
       pullDocumentFromDynamoStrategy: new PullDocumentByCodeStrategy(
         req,
@@ -159,5 +158,33 @@ const getStrategies = (
   }
   return strategies;
 };
+
+function createClientMetadata(redirect_uri, req, config) {
+  let clientMetadata = {
+    redirect_uris: [redirect_uri],
+  };
+
+  const basicAuth = parseBasicAuth(req);
+  if (basicAuth) {
+    clientMetadata.client_id = basicAuth.username;
+    clientMetadata.client_secret = basicAuth.password;
+  } else if (req.body.client_id && req.body.client_secret) {
+    clientMetadata.client_id = req.body.client_id;
+    clientMetadata.client_secret = req.body.client_secret;
+    delete req.body.client_id;
+    delete req.body.client_secret;
+  } else if (config.enable_pkce_authorization_flow && req.body.client_id) {
+    clientMetadata.token_endpoint_auth_method = "none";
+    clientMetadata.client_id = req.body.client_id;
+    delete req.body.client_id;
+  } else {
+    throw {
+      status: 401,
+      error: "invalid_client",
+      error_description: "Client authentication failed",
+    };
+  }
+  return clientMetadata;
+}
 
 module.exports = { buildTokenHandlerClient };
